dis.rs: Add test against a dump of the two-board ALU ROM.
It validated once we discovered the ALU (74LS181) S0-4 bits were reversed
from the ordering in the one-board's ALU instruction.

Well, _mostly_ validated--addr 4 of the ROM (ROL) is incorrect, but
interestingly has a note on the printout handwritten in Bill's handwriting
that says simply, "wrong". So I'm going to assume that hand-written note is
correct and that the ROM _should_ be something else.

It also appears that ROL was never used in the two-board microcode so the
fact that it was wrong never mattered...