First pass at assembler. I haven't double checked all the output yet, just what is in the tests.
I had to add JumpDest::Myself, which now is a little redundant with the
"JumpDest::Absolute(addr) if addr == addr" stuff in the disassembler. Maybe
have the disassembler figure that out earler?

I found some crates to do ihex and srecords, so I added a --format option.

I figured it would easier for me to do the de-interleaving here, so I added
a --interleave option if you want to split the assembled rom image into high
and low byte sections. It's not symmetric--the disassembler doesn't support
interleaved rom images (...yet?).

I'm not super happy with how many lines of code it takes to dump images and
deal with interleaving and stdout/file writing. Rust made parts of it
tough:(why is force adding an extention (and not replacing it) so darn
hard??? Notice calling the ihex/srecord libraries is shorter code. Sigh.).

I don't like SymbolTable living in dis.rs. Where should it move though?
Should it even be called "SymbolTable"? All that exists there right now are
labels, ie, jump destinations.

I implemented parsing things to put in the symbol table:

    DAVID = 172

...but don't do anything with it. Where would that be valid? The only place
a number could go would be MV immediate. Would it be:

    MV #DAVID,GP10

or

    MV DAVID,GP10

?

Also, GPs... Should they be named? Should their names be hardcoded to what
the firmware used? Do we even know what the firmware uses? We only have a
listing of the old code and it might have changed.

Ideally I'd like something like:

    DAVID = GP4

and then use it in MV source and dests:

    MV #27,DAVID
    MV DAVID,A

How to represent that? My symbol table stuff currently expects numbers. Is
there _anything_ other than GPs we'd choose a name for?? Should there be
something specific for them? Should it be a general macro system? If so,
does _every_ parser have to have a special case to look up the operand and
replace it with something else? There's no central way to do that at the
moment...

I love the way emit() is so compact. I tried to let the format of the
instructions shine through with all the raw binary. But I used some consts
from cdf8.rs that dis.rs was using... Is it better to have them share to cut
down on duplication and bugs, or have them raw to make the code super super
clear?